+++ whatsnew.md
@@ -1,3 +1,11 @@
+## 2.5.1
+
+### Fixes
+
+-   Impossible to arrange first level items as a column
+-   Impossible to fix certain items as always visible
+-   Collapsed items are still visible outside container on overflow
+
 ## 2.5.0
 
 ### Initial features